requirements.txt
streamlit==1.32.0
pandas==2.0.3
scikit-learn==1.4.0
xgboost==2.0.3
matplotlib==3.8.0
seaborn==0.13.0
joblib==1.3.2
Pillow==10.2.0